pyproject.toml
[tool.poetry]
name = "hello_world"
version = "0.1.0"
description = "Microservices auction system using RabbitMQ"
authors = ["Eduardo Soviersovski",  "Pedro Tortola"]
readme = "README.md"
packages = [{ include = "hello_world" }]

[tool.poetry.dependencies]
python = "^3.12"
pika = "1.0.0"
requests = "2.31.0"
pycryptodome = "^3.23.0"
cryptography = "^45.0.6"

[build-system]
requires = ["poetry-core>=1.4.0"]
build-backend = "poetry.core.masonry.api"
